+++ ton_client/src/json_interface/mod.rs
@@ -18,6 +18,7 @@ pub(crate) mod handlers;
 pub(crate) mod interop;
 pub(crate) mod net;
 pub(crate) mod processing;
+pub(crate) mod utils;
 
 pub(crate) mod modules;
 mod registrar;